Apply structure
