+++ app/components/face/face.jsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import { Component } from 'react';
 
-import randomColor from 'randomcolor';
 import './face.styl';
 
 /**
@@ -13,12 +12,11 @@ export default class Face extends Component {
 
     constructor(props) {
         super(props);
-        this.color = randomColor();
     }
 
     render() {
         const style = {
-            background: this.color,
+            background: this.props.color,
             transform: this.props.transform,
             width: this.props.width,
             height: this.props.height
