+++ app/utils/logger.jsx
@@ -1,0 +1,3 @@
+export default function logger(store) {
+    console.log(store);
+};
